DAT-172 update functions and check permissions with token

+++ server/src/main/java/datavault/server/controllers/FilesController.java
@@ -4,6 +4,7 @@ import datavault.server.dto.FileGetDTO;
 import datavault.server.dto.FilePostDTO;
 import datavault.server.dto.FilePutDTO;
 import datavault.server.services.FilesService;
+import datavault.server.services.PermissionService;
 import jakarta.websocket.server.PathParam;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -14,9 +15,10 @@ import java.util.List;
 @RestController
 @RequestMapping("file")
 public class FilesController {
-
     @Autowired
     FilesService filesService;
+    @Autowired
+    PermissionService permissionService;
 
     @PostMapping
     public ResponseEntity<String> register(@RequestBody FilePostDTO filePostDTO) {
@@ -32,21 +34,25 @@ public class FilesController {
 
     @GetMapping
     public ResponseEntity<List<FileGetDTO>> getAllFiles() {
+        permissionService.validateAdmin();
         return ResponseEntity.ok(filesService.getAll());
     }
 
     @GetMapping("/{username}")
     public ResponseEntity<List<FileGetDTO>> getAllFilesByUser(@PathVariable("username") String username) {
+        permissionService.validateAccessToDataOfUser(username);
         return ResponseEntity.ok(filesService.getAllByUsername(username));
     }
 
     @GetMapping("/shared/{username}")
     public ResponseEntity<List<FileGetDTO>> getAllMyFiles(@PathVariable("username") String username) {
+        permissionService.validateAccessToDataOfUser(username);
         return ResponseEntity.ok(filesService.getMyAvailableFiles(username));
     }
 
     @DeleteMapping("/{fileId}")
     public ResponseEntity<?> removeFile(@PathVariable("fileId") String fileId) {
+        permissionService.validateMenagePermissionForFile(fileId);
         filesService.removeFile(fileId);
         return ResponseEntity.ok().build();
     }
